etl/pyproject.toml
[tool.poetry]
name = "trader-etl-service"
version = "0.1.0"
description = "Permanently running ETL service for experimental trader project"
authors = ["Alex Bausk <bauskas@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.10.3"
google-cloud-bigquery = "^1.16"
requests = "^2.22"
python-dateutil = "^2.8"
gunicorn = "^19.9"
aiohttp = "^3.5"
cchardet = "^2.1"
aiodns = "^2.0"

[tool.poetry.dev-dependencies]
ptvsd = "^4.2"
pydevd-pycharm = "192.5587.24"
pylint = "^2.3"
rope = "^0.14.0"
autopep8 = "^1.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
